Balancing

+++ DissertationProject/Assets/Scripts/Enemy.cs
@@ -9,7 +9,7 @@ public class Enemy : MonoBehaviour
     int pathNodeIndex = 0;
     public float speed = 5;
     Vector2 direction;
-    public int health = 5;
+    public float health = 5.0f;
     ScoreManager scoreManager;
     // Start is called before the first frame update
     void Start()
